[fix] NodeBird 게시글 삭제 기능 수정
- 게시물 삭제 요청 시, 이벤트가 일어나는 포스트가 아닌 다른 포스트가 삭제되는 이슈 해결 (쿼리 셀렉터 경로 문제)
- 게시물 작성자 ID 와 게시물 삭제 요청자 ID가 일치하는 경우에 대해서만 삭제가 이뤄지도록 처리

+++ section8/routes/post.js
@@ -3,7 +3,7 @@ const multer = require('multer');
 const path = require('path');
 const fs = require('fs');
 
-const { Post, Hashtag } = require('../models');
+const { User, Post, Hashtag } = require('../models');
 const { isLoggedIn } = require('./middlewares');
 
 const router = express.Router()
@@ -35,10 +35,14 @@ router.post('/img', isLoggedIn, upload.single('img'), (req, res) => {
 
 router.delete('/:twitId/delete', isLoggedIn, async (req, res, next) => {
   try {
-    const post = Post.findOne({ where : { id : req.params.twitId }});
+    const post = await Post.findOne({ where : { id : req.params.twitId }});
     if (post) {
-      await Post.destroy({where : {id : req.params.twitId}});
-      res.send('success');
+      if (post.UserId == req.user.id){
+        await Post.destroy({where : {id : req.params.twitId}});
+        res.send('success');
+      } else {
+        res.status(404).send('No permissions');
+      }
     } else {
       res.status(404).send('No twit');
     }
